+++ src/States/GameState.cxx
@@ -5,8 +5,6 @@ void GameState::initMap()
 {
     map = std::make_unique<Map>("My New World", 56465456464654, data.activeResourcePack->tileDB,
                                 data.activeResourcePack->textures.at("TileSheet"), data.gridSize, data.scale);
-
-    // map->loadRegion({6, 4});
 }
 
 void GameState::initThisPlayer()
@@ -101,23 +99,21 @@ void GameState::updatePlayerCamera()
 
     if ((playerCamera.getCenter().x - playerCamera.getSize().x / 2.f) <= 0.f)
     {
-        playerCamera.setCenter(
-            sf::Vector2f(static_cast<int>(playerCamera.getSize().x / 2.f), playerCamera.getCenter().y));
+        playerCamera.setCenter(sf::Vector2f((playerCamera.getSize().x / 2.f), playerCamera.getCenter().y));
     }
     else if ((playerCamera.getCenter().x + playerCamera.getSize().x / 2.f) >= map->getRealDimensions().x)
     {
-        playerCamera.setCenter(sf::Vector2f(
-            static_cast<int>(map->getRealDimensions().x - playerCamera.getSize().x / 2.f), playerCamera.getCenter().y));
+        playerCamera.setCenter(
+            sf::Vector2f((map->getRealDimensions().x - playerCamera.getSize().x / 2.f), playerCamera.getCenter().y));
     }
     if ((playerCamera.getCenter().y - playerCamera.getSize().y / 2.f) <= 0.f)
     {
-        playerCamera.setCenter(
-            sf::Vector2f(playerCamera.getCenter().x, static_cast<int>(playerCamera.getSize().y / 2.f)));
+        playerCamera.setCenter(sf::Vector2f(playerCamera.getCenter().x, (playerCamera.getSize().y / 2.f)));
     }
     else if ((playerCamera.getCenter().y + playerCamera.getSize().y / 2.f) >= map->getRealDimensions().y)
     {
-        playerCamera.setCenter(sf::Vector2f(
-            playerCamera.getCenter().x, static_cast<int>(map->getRealDimensions().y - playerCamera.getSize().y / 2.f)));
+        playerCamera.setCenter(
+            sf::Vector2f(playerCamera.getCenter().x, (map->getRealDimensions().y - playerCamera.getSize().y / 2.f)));
     }
 }
